+++ lib/groupdate.rb
@@ -1,27 +1,7 @@
+require "active_support/time"
 require "groupdate/version"
-require "groupdate/scopes"
 require "groupdate/enumerable"
-
-ActiveRecord::Base.send :extend, Groupdate::Scopes
-
-# hack for **unfixed** rails issue
-# https://github.com/rails/rails/issues/7121
-module ActiveRecord
-  module Calculations
-
-    private
-
-    def column_alias_for_with_hack(*keys)
-      if keys.first.is_a?(Groupdate::OrderHack)
-        keys.first.field
-      else
-        column_alias_for_without_hack(*keys)
-      end
-    end
-    alias_method_chain :column_alias_for, :hack
-
-  end
-end
+require "groupdate/active_record"
 
 module Groupdate
   mattr_accessor :week_start, :day_start, :time_zone
