+++ lib/groupdate/active_record.rb
@@ -1,0 +1,23 @@
+require "active_record"
+require "groupdate/scopes"
+
+ActiveRecord::Base.send(:extend, Groupdate::Scopes)
+
+# hack for **unfixed** rails issue
+# https://github.com/rails/rails/issues/7121
+module ActiveRecord
+  module Calculations
+
+    private
+
+    def column_alias_for_with_hack(*keys)
+      if keys.first.is_a?(Groupdate::OrderHack)
+        keys.first.field
+      else
+        column_alias_for_without_hack(*keys)
+      end
+    end
+    alias_method_chain :column_alias_for, :hack
+
+  end
+end
